feat: 인게임 대기방 hook 분리
